GLSP-974: Use interface-injection for all subclasses of ModelState
refs https://github.com/eclipse-glsp/glsp/issues/974
