+++ plugins/org.eclipse.emfcloud.modelserver.glsp.integration/src/org/eclipse/emfcloud/modelserver/glsp/EMSModelState.java
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2021-2023 EclipseSource and others.
+ * Copyright (c) 2023 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -12,52 +12,13 @@ package org.eclipse.emfcloud.modelserver.glsp;
 
 import java.util.Optional;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.eclipse.emf.ecore.EObject;
-import org.eclipse.glsp.graph.GModelIndex;
-import org.eclipse.glsp.graph.GModelRoot;
-import org.eclipse.glsp.server.emf.EMFModelIndex;
 import org.eclipse.glsp.server.emf.EMFModelState;
-import org.eclipse.glsp.server.emf.notation.EMFSemanticIdConverter;
-import org.eclipse.glsp.server.session.ClientSession;
 
-import com.google.inject.Inject;
+public interface EMSModelState extends EMFModelState {
+   void setSemanticModel(final EObject semanticModel);
 
-public class EMSModelState extends EMFModelState {
+   EObject getSemanticModel();
 
-   private static final Logger LOGGER = LogManager.getLogger(EMSModelState.class);
-
-   protected EObject semanticModel;
-
-   @Inject
-   protected EMSModelServerAccess modelServerAccess;
-   @Inject
-   protected EMFSemanticIdConverter semanticIdConverter;
-
-   @Override
-   protected GModelIndex getOrUpdateIndex(final GModelRoot newRoot) {
-      return EMFModelIndex.getOrCreate(getRoot(), semanticIdConverter);
-   }
-
-   public void setSemanticModel(final EObject semanticModel) { this.semanticModel = semanticModel; }
-
-   public EObject getSemanticModel() { return this.semanticModel; }
-
-   public <T extends EObject> Optional<T> getSemanticModel(final Class<T> clazz) {
-      return Optional.ofNullable(this.semanticModel).filter(clazz::isInstance).map(clazz::cast);
-   }
-
-   @Override
-   public void sessionDisposed(final ClientSession clientSession) {
-      if (this.semanticModel != null) {
-         this.semanticModel = null;
-         modelServerAccess.close().thenAccept(response -> {
-            LOGGER.warn("Error on disposing ClientSession: " + response.getMessage());
-         });
-         modelServerAccess.unsubscribe();
-      }
-      super.sessionDisposed(clientSession);
-   }
-
+   <T extends EObject> Optional<T> getSemanticModel(final Class<T> clazz);
 }
